+++ src/RhinoInside.Revit.GH/Components/Element/Parameters.cs
@@ -425,13 +425,13 @@ namespace RhinoInside.Revit.GH.Components
     }
   }
 
-  public class ElementParameters : Component
+  public class QueryElementParameters : Component
   {
     public override Guid ComponentGuid => new Guid("44515A6B-84EE-4DBD-8241-17EDBE07C5B6");
     public override GH_Exposure Exposure => GH_Exposure.quarternary;
 
-    public ElementParameters()
-    : base("Element Parameters", "Parameters", "Get the parameters of the specified Element", "Revit", "Element")
+    public QueryElementParameters()
+    : base("Query Element Parameters", "Parameters", "Get the parameters of the specified Element", "Revit", "Element")
     { }
 
     protected override void RegisterInputParams(GH_InputParamManager manager)